convert PlayerContainer to TS
