+++ src/interfaces/IPlayerContainerProps.ts
@@ -1,0 +1,15 @@
+
+export interface IPlayerContainerProps {
+  dealerHasControl: boolean;
+  gameStatus: number;
+  gameStatusFlag: boolean;
+  isCardDescVisible: boolean;
+  isDealerHandVisible: boolean;
+  isDeckCalloutVisible: boolean;
+  isHandValueVisible: boolean;
+  player: any;
+  playerHand: any;
+  playerKey: string;
+  playerStats: any;
+  minimumBet: number;
+}
